:recycle:  refactor: common 모듈 implements 및 OptionalLong 오류 해결 [#34]
