+++ waggle-api/src/main/java/com/trip/api/chatting/dto/request/CreateChatMessageRequest.java
@@ -1,7 +1,7 @@
 package com.trip.api.chatting.dto.request;
 
-import com.trip.api.config.ValidateType;
-import com.trip.api.type.MessageType;
+import com.trip.common.annotation.ValidateType;
+import com.trip.common.type.MessageType;
 import lombok.Getter;
 
 import javax.validation.constraints.NotBlank;
