+++ waggle-api/src/main/java/com/trip/api/config/QueryDslConfig.java
@@ -1,4 +1,4 @@
-package com.trip.tripinworld.global.config;
+package com.trip.api.config;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
